+++ resources/assets/components/pages/CampaignPage/campaign-page.scss
@@ -1,6 +1,6 @@
 @import '../../../scss/next-toolbox';
 
-.campaign-subpage {
+.campaign-page {
   font-family: $primary-font-family;
   font-size: $font-regular;
   line-height: 1.4;
